// src/app/models/IBaseCard..ts
export interface IBaseCard {

	id: string;

	manufacturer: string;

	type: string;

	country: string;

	body: string;

	driveWheel: string;

	fuelType: string;

	abs: string;

	tractionControl: string;

	imageUrl: string;

	logoURL: string;

	carPageUrl: string;

	objectPositionHorizontal: string;

	objectPositionVertical: string;

	objectWidth: string;

	objectHeight: string;

	year: number;

	doors: number;

	cornering: number;

	seats: number;

	fuelTankCapacity: number;

	engineCapacity: number;

	powerKW: number;

	powerHP: number;

	maxTorque: number;

	topSpeed: number;

	weight: number;

	length: number;

	width: number;

	height: number;

	groundClearance: number;

	acceleration: number;

	gear1st: number;

	gear2nd: number;

	gear3rd: number;

	gear4th: number;

	gear5th: number;

	gear6th: number;

	finalDrive: number;
}